+++ NEWS.md
@@ -1,3 +1,16 @@
+# skytrackr 1.1
+
+Where release 1.0 was tuned for use on the common swift (an ideal light logger species)
+this release sees improvements in which cover expanding the methodology to
+species which have less than optimal (noisy) light logger data. It provides new
+methods to correctly estimate parameter ranges (attenuation factors due to
+preferred secluded habitats).
+
+- introduction of a filter function `stk_filter()`, but helpful in visualizing data
+- introduction of the `stk_calibrate()` function to estimate scale factors for optimization
+- correction to the subsetting of the data (see `stk_filter()`) to accommodate noisy data
+- trap rendering error on intermediate plots, which corrupts optimization
+
 # skytrackr 1.0
 
 - removed seed from `stk_cluster()`